feat: add service logic
