+++ handler/webhook.go
@@ -3,45 +3,18 @@ package handler
 import (
 	"net/http"
 
+	"github.com/forjadev/gun-organization/service"
 	"github.com/gin-gonic/gin"
 )
 
-type GithubWebhookUserManagePayload struct {
-	Action string `json:"action"`
-	Scope  string `json:"scope"`
-
-	Member struct {
-		ID    uint   `json:"id"`
-		Login string `json:"login"`
-		Type  string `json:"type"`
-		Url   string `json:"html_url"`
-	}
-
-	Sender struct {
-		ID    uint   `json:"id"`
-		Login string `json:"login"`
-		Type  string `json:"type"`
-	}
-
-	Team struct {
-		ID   uint   `json:"id"`
-		Name string `json:"name"`
-		Slug string `json:"slug"`
-	}
-
-	Organization struct {
-		ID    uint   `json:"id"`
-		Login string `json:"login"`
-		URL   string `json:"url"`
-	}
-}
-
 func WebhookReceiverHandle(c *gin.Context) {
-	var postData GithubWebhookUserManagePayload
+	postData := new(service.GithubWebhookUserManagePayload)
 	if err := c.ShouldBindJSON(&postData); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
+
+	service.WebhookHandler(c, postData)
 
 	c.JSON(http.StatusOK, gin.H{"data": postData})
 }
